+++ backend/app/services/ai_service.py
@@ -11,16 +11,19 @@ class AIService:
     def __init__(self):
         self.model = genai.GenerativeModel("gemini-2.5-flash")
 
-    def get_response(self, message: str):
-
-        prompt = f"""
-{SYSTEM_PROMPT}
+    def get_response(self, history: list):
+            conversation = ""
+            for msg in history:#a loop that converts the list -> string
+                 conversation += f"{msg['role'].capitalize()}: {msg['content']}\n"
+                 
+            prompt = f"""
+            {SYSTEM_PROMPT}
         
-User:
-{message}
+            Conversation:
+            {conversation}
 
-Respond as DevMind AI.
-        """
+            Respond as DevMind AI.
+            """
 
-        response = self.model.generate_content(prompt)
-        return response.text
+            response = self.model.generate_content(prompt)
+            return response.text